% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{DEFAULT_target_lod}
\alias{DEFAULT_target_lod}
\title{Default Limits of Detection (LOD)}
\usage{
DEFAULT_target_lod
}
\description{
A list containing the default upper (Hi) and lower (Lo) limits of detection
for various qPCR targets.
}
