% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qpcrpipeline-package.R
\name{qpcrpipeline}
\alias{qpcrpipeline}
\title{qpcrpipeline: MIQE-aligned qPCR Data Cleaning and Consolidation Pipeline}
\description{
An R package for cleaning, auditing, and consolidating qPCR data.
}
