% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{use_qpcr_template}
\alias{use_qpcr_template}
\title{Create a new qPCR project template}
\usage{
use_qpcr_template(path = ".")
}
\arguments{
\item{path}{Target directory path where the template should be created.}
}
\value{
Invisible TRUE on success.
}
\description{
Copies the example runner script and folder structure to your local directory.
}
